ryno-v2-scripts/requirements.txt
fastapi==0.95.0
langchain==0.0.131
numpy==1.23.5
openai==0.27.0
pdfplumber==0.8.0
pinecone_client==2.2.1
protobuf==4.22.1
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.2
uvicorn==0.21.1
google-cloud-storage
